+++ packages/rendermime/src/index.ts
@@ -5,9 +5,10 @@
 import '../style/index.css';  // Why is this first?
 
 export * from '@jupyterlab/rendermime-interfaces';
+export * from './factories';
 export * from './latex';
 export * from './mimemodel';
 export * from './outputmodel';
+export * from './renderers';
 export * from './rendermime';
-export * from './factories';
 export * from './widgets';
