+++ packages/rendermime/src/widgets.ts
@@ -3,6 +3,10 @@
 | Distributed under the terms of the Modified BSD License.
 |----------------------------------------------------------------------------*/
 import {
+  IRenderMime
+} from '@jupyterlab/rendermime-interfaces';
+
+import {
   Message
 } from '@phosphor/messaging';
 
@@ -11,16 +15,11 @@ import {
 } from '@phosphor/widgets';
 
 import {
-  IRenderMime
-} from '@jupyterlab/rendermime-interfaces';
-
-import {
   typeset
 } from './latex';
 
-import {
-  RenderHelpers
-} from './renderhelpers';
+import * as renderers
+  from './renderers';
 
 
 /**
@@ -72,8 +71,8 @@ abstract class RenderedCommon extends Widget implements IRenderMime.IRenderer {
   renderModel(model: IRenderMime.IMimeModel): Promise<void> {
     // TODO compare model against old model for early bail?
 
-    // Set the trusted flag on the node.
-    this.node.dataset['trusted'] = `${model.trusted}`;
+    // Toggle the trusted class on the widget.
+    this.toggleClass('jp-mod-trusted', model.trusted);
 
     // Render the actual content.
     return this.render(model);
@@ -130,51 +129,8 @@ class RenderedHTML extends RenderedHTMLCommon {
    * @returns A promise which resolves when rendering is complete.
    */
   render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderHTML({
-      node: this.node,
-      source: String(model.data[this.mimeType]),
-      trusted: model.trusted,
-      resolver: this.resolver,
-      sanitizer: this.sanitizer,
-      linkHandler: this.linkHandler,
-      shouldTypeset: this.isAttached
-    });
-  }
-
-  /**
-   * A message handler invoked on an `'after-attach'` message.
-   */
-  onAfterAttach(msg: Message): void {
-    typeset(this.node);
-  }
-}
-
-
-/**
- * A mime renderer for displaying Markdown with embeded latex.
- */
-export
-class RenderedMarkdown extends RenderedHTMLCommon {
-  /**
-   * Construct a new rendered markdown widget.
-   *
-   * @param options - The options for initializing the widget.
-   */
-  constructor(options: IRenderMime.IRendererOptions) {
-    super(options);
-    this.addClass('jp-RenderedMarkdown');
-  }
-
-  /**
-   * Render a mime model.
-   *
-   * @param model - The mime model to render.
-   *
-   * @returns A promise which resolves when rendering is complete.
-   */
-  render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderMarkdown({
-      node: this.node,
+    return renderers.renderHTML({
+      host: this.node,
       source: String(model.data[this.mimeType]),
       trusted: model.trusted,
       resolver: this.resolver,
@@ -216,8 +172,8 @@ class RenderedLatex extends RenderedCommon {
    * @returns A promise which resolves when rendering is complete.
    */
   render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderLatex({
-      node: this.node,
+    return renderers.renderLatex({
+      host: this.node,
       source: String(model.data[this.mimeType]),
       shouldTypeset: this.isAttached
     });
@@ -245,7 +201,6 @@ class RenderedImage extends RenderedCommon {
   constructor(options: IRenderMime.IRendererOptions) {
     super(options);
     this.addClass('jp-RenderedImage');
-    this.node.appendChild(document.createElement('img'));
   }
 
   /**
@@ -256,29 +211,30 @@ class RenderedImage extends RenderedCommon {
    * @returns A promise which resolves when rendering is complete.
    */
   render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderImage({
-      model,
+    return renderers.renderImage({
+      host: this.node,
       mimeType: this.mimeType,
-      node: this.node.firstChild as HTMLImageElement
+      source: String(model.data[this.mimeType]),
+      width: model.metadata.width as number | undefined,
+      height: model.metadata.height as number | undefined
     });
   }
 }
 
 
 /**
- * A widget for displaying plain text and console text.
+ * A mime renderer for displaying Markdown with embeded latex.
  */
 export
-class RenderedText extends RenderedCommon {
+class RenderedMarkdown extends RenderedHTMLCommon {
   /**
-   * Construct a new rendered text widget.
+   * Construct a new rendered markdown widget.
    *
    * @param options - The options for initializing the widget.
    */
   constructor(options: IRenderMime.IRendererOptions) {
     super(options);
-    this.addClass('jp-RenderedText');
-    this.node.appendChild(document.createElement('pre'));
+    this.addClass('jp-RenderedMarkdown');
   }
 
   /**
@@ -289,75 +245,12 @@ class RenderedText extends RenderedCommon {
    * @returns A promise which resolves when rendering is complete.
    */
   render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderText({
-      model,
-      mimeType: this.mimeType,
-      node: this.node.firstChild as HTMLElement
-    });
-  }
-}
-
-
-/**
- * A widget for displaying/executing JavaScript.
- */
-export
-class RenderedJavaScript extends RenderedCommon {
-  /**
-   * Construct a new rendered Javascript widget.
-   *
-   * @param options - The options for initializing the widget.
-   */
-  constructor(options: IRenderMime.IRendererOptions) {
-    super(options);
-    this.addClass('jp-RenderedJavaScript');
-  }
-
-  /**
-   * Render a mime model.
-   *
-   * @param model - The mime model to render.
-   *
-   * @returns A promise which resolves when rendering is complete.
-   */
-  render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderJavaScript({
-      model,
-      node: this.node,
-      mimeType: this.mimeType
-    });
-  }
-}
-
-
-/**
- * A widget for displaying SVG content.
- */
-export
-class RenderedSVG extends RenderedCommon {
-  /**
-   * Construct a new rendered SVG widget.
-   *
-   * @param options - The options for initializing the widget.
-   */
-  constructor(options: IRenderMime.IRendererOptions) {
-    super(options);
-    this.addClass('jp-RenderedSVG');
-  }
-
-  /**
-   * Render a mime model.
-   *
-   * @param model - The mime model to render.
-   *
-   * @returns A promise which resolves when rendering is complete.
-   */
-  render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderSVG({
-      model,
-      node: this.node,
-      mimeType: this.mimeType,
+    return renderers.renderMarkdown({
+      host: this.node,
+      source: String(model.data[this.mimeType]),
+      trusted: model.trusted,
       resolver: this.resolver,
+      sanitizer: this.sanitizer,
       linkHandler: this.linkHandler,
       shouldTypeset: this.isAttached
     });
@@ -391,10 +284,83 @@ class RenderedPDF extends RenderedCommon {
    * Render a mime model.
    */
   render(model: IRenderMime.IMimeModel): Promise<void> {
-    return RenderHelpers.renderPDF({
-      model,
-      node: this.node,
-      mimeType: this.mimeType
+    return renderers.renderPDF({
+      host: this.node,
+      source: String(model.data[this.mimeType]),
+      trusted: model.trusted
+    });
+  }
+}
+
+
+/**
+ * A widget for displaying SVG content.
+ */
+export
+class RenderedSVG extends RenderedCommon {
+  /**
+   * Construct a new rendered SVG widget.
+   *
+   * @param options - The options for initializing the widget.
+   */
+  constructor(options: IRenderMime.IRendererOptions) {
+    super(options);
+    this.addClass('jp-RenderedSVG');
+  }
+
+  /**
+   * Render a mime model.
+   *
+   * @param model - The mime model to render.
+   *
+   * @returns A promise which resolves when rendering is complete.
+   */
+  render(model: IRenderMime.IMimeModel): Promise<void> {
+    return renderers.renderSVG({
+      host: this.node,
+      source: String(model.data[this.mimeType]),
+      trusted: model.trusted,
+      resolver: this.resolver,
+      linkHandler: this.linkHandler,
+      shouldTypeset: this.isAttached
+    });
+  }
+
+  /**
+   * A message handler invoked on an `'after-attach'` message.
+   */
+  onAfterAttach(msg: Message): void {
+    typeset(this.node);
+  }
+}
+
+
+/**
+ * A widget for displaying plain text and console text.
+ */
+export
+class RenderedText extends RenderedCommon {
+  /**
+   * Construct a new rendered text widget.
+   *
+   * @param options - The options for initializing the widget.
+   */
+  constructor(options: IRenderMime.IRendererOptions) {
+    super(options);
+    this.addClass('jp-RenderedText');
+  }
+
+  /**
+   * Render a mime model.
+   *
+   * @param model - The mime model to render.
+   *
+   * @returns A promise which resolves when rendering is complete.
+   */
+  render(model: IRenderMime.IMimeModel): Promise<void> {
+    return renderers.renderText({
+      host: this.node,
+      source: String(model.data[this.mimeType])
     });
   }
 }
